+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.9.0] - 2021-10-25
 ### Fixed
 - query: support borrowed long strings, [#22](https://github.com/loyd/clickhouse.rs/issues/22).
 - query: read the whole response of DDL queries.
@@ -128,7 +130,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client::insert()` for inserting into tables.
 - `Client::query()` for selecting from tables and DDL statements.
 
-[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.8.1...HEAD
+[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/loyd/clickhouse.rs/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/loyd/clickhouse.rs/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/loyd/clickhouse.rs/compare/v0.7.2...v0.8.0
 [0.7.2]: https://github.com/loyd/clickhouse.rs/compare/v0.7.1...v0.7.2